cheri: Apply automatic FixCapFormat clang-tidy changes
Apply a first round of FixCapFormat clang-tidy changes.
There will be more changes once we change the type of
more fields and variables to uintptr_t.

+++ drivers/net/ethernet/cadence/macb_main.c
@@ -5614,7 +5614,7 @@ static int macb_probe(struct platform_device *pdev)
 
 	netdev_info(dev, "Cadence %s rev 0x%08x at 0x%08lx irq %d (%pM)\n",
 		    macb_is_gem(bp) ? "GEM" : "MACB", macb_readl(bp, MID),
-		    dev->base_addr, dev->irq, dev->dev_addr);
+		    (unsigned long)dev->base_addr, dev->irq, dev->dev_addr);
 
 	pm_runtime_mark_last_busy(&bp->pdev->dev);
 	pm_runtime_put_autosuspend(&bp->pdev->dev);
